tools/requirements.txt
requests
jinja2
python-youtube
google-auth-oauthlib
google-auth-httplib2
google-api-python-client